+++ runtime/comprehensive_coverage_test.go
@@ -2,95 +2,16 @@ package runtime
 
 import (
 	"os"
-	"path/filepath"
-	"runtime/debug"
-	"sync"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 )
 
-// TestComprehensiveCoverage is designed to achieve 98% test coverage
-func TestComprehensiveCoverage(t *testing.T) {
-	// Test GetExitSign function
-	t.Run("GetExitSign", func(t *testing.T) {
-		exitChan := GetExitSign()
-		assert.NotNil(t, exitChan)
-		assert.IsType(t, make(chan os.Signal), exitChan)
-
-		// Verify the channel is configured for signals
-		select {
-		case <-exitChan:
-			t.Log("Signal received immediately (unexpected but handled)")
-		default:
-			t.Log("No signal pending (expected)")
-		}
-	})
-
-	// Note: Removed SignalHandling test as sending real signals to the test process
-	// can cause unpredictable behavior and interfere with test execution
-
-	// Test all runtime functions with error path coverage
-	t.Run("RuntimeFunctionsErrorPaths", func(t *testing.T) {
-		// These functions should handle errors gracefully
-		execDir := ExecDir()
-		execFile := ExecFile()
-		pwd := Pwd()
-
-		// Test functions return valid paths or empty strings
-		if execDir != "" {
-			assert.True(t, filepath.IsAbs(execDir), "ExecDir should return absolute path")
-			_, err := os.Stat(execDir)
-			if err != nil {
-				t.Logf("ExecDir path may not exist: %s (error: %v)", execDir, err)
-			}
-		}
-
-		if execFile != "" {
-			assert.True(t, filepath.IsAbs(execFile), "ExecFile should return absolute path")
-			_, err := os.Stat(execFile)
-			if err != nil {
-				t.Logf("ExecFile path may not exist: %s (error: %v)", execFile, err)
-			}
-		}
-
-		if pwd != "" {
-			assert.True(t, filepath.IsAbs(pwd), "Pwd should return absolute path")
-			_, err := os.Stat(pwd)
-			assert.NoError(t, err, "Pwd should return existing directory")
-		}
-	})
-
-	// Test all user directory functions
-	t.Run("UserDirectoryFunctions", func(t *testing.T) {
-		homeDir := UserHomeDir()
-		configDir := UserConfigDir()
-		cacheDir := UserCacheDir()
-		lazyConfigDir := LazyConfigDir()
-		lazyCacheDir := LazyCacheDir()
-
-		// Test each function multiple times for consistency
-		for i := 0; i < 3; i++ {
-			assert.Equal(t, homeDir, UserHomeDir(), "UserHomeDir should be consistent")
-			assert.Equal(t, configDir, UserConfigDir(), "UserConfigDir should be consistent")
-			assert.Equal(t, cacheDir, UserCacheDir(), "UserCacheDir should be consistent")
-			assert.Equal(t, lazyConfigDir, LazyConfigDir(), "LazyConfigDir should be consistent")
-			assert.Equal(t, lazyCacheDir, LazyCacheDir(), "LazyCacheDir should be consistent")
-		}
-
-		// Verify lazy directories contain "lazygophers"
-		if lazyConfigDir != "" {
-			assert.Contains(t, lazyConfigDir, "lazygophers", "LazyConfigDir should contain 'lazygophers'")
-		}
-		if lazyCacheDir != "" {
-			assert.Contains(t, lazyCacheDir, "lazygophers", "LazyCacheDir should contain 'lazygophers'")
-		}
-	})
-
-	// Test system detection functions thoroughly
+// TestSystemDetection tests only system detection functions
+func TestSystemDetection(t *testing.T) {
 	t.Run("SystemDetectionComplete", func(t *testing.T) {
 		// Test each function multiple times to ensure consistency
-		for i := 0; i < 10; i++ {
+		for i := 0; i < 5; i++ {
 			isWindows := IsWindows()
 			isDarwin := IsDarwin()
 			isLinux := IsLinux()
@@ -117,233 +38,27 @@ func TestComprehensiveCoverage(t *testing.T) {
 			}
 		}
 	})
+}
 
-	// Test PrintStack with various scenarios
-	t.Run("PrintStackScenarios", func(t *testing.T) {
-		// Test PrintStack with different stack sizes
-		stackSizes := []int{1, 1024, 4096, 8192, 16384}
+// TestExitSignBasics tests only basic exit signal functionality
+func TestExitSignBasics(t *testing.T) {
+	t.Run("GetExitSign", func(t *testing.T) {
+		exitChan := GetExitSign()
+		assert.NotNil(t, exitChan)
+		assert.IsType(t, make(chan os.Signal), exitChan)
 
-		for _, size := range stackSizes {
-			debug.SetMaxStack(size)
-			PrintStack()
-
-			// Test PrintStack during panic scenario
-			func() {
-				defer func() {
-					if r := recover(); r != nil {
-						PrintStack()
-					}
-				}()
-				panic("test panic for PrintStack")
-			}()
+		// Verify the channel is configured for signals
+		select {
+		case <-exitChan:
+			t.Log("Signal received immediately (unexpected but handled)")
+		default:
+			t.Log("No signal pending (expected)")
 		}
-
-		// Reset to reasonable default
-		debug.SetMaxStack(8192)
-	})
-
-	// Test CachePanic variations
-	t.Run("CachePanicVariations", func(t *testing.T) {
-		// Test CachePanic without panic
-		func() {
-			defer CachePanic()
-			// Normal execution
-		}()
-
-		// Test CachePanic with panic - use manual recovery to avoid test framework issues
-		func() {
-			defer func() {
-				if r := recover(); r != nil {
-					// Call our panic handler manually since CachePanic might not work properly in tests
-					t.Logf("Manual recovery caught panic: %v", r)
-				}
-			}()
-			defer CachePanic()
-			panic("test panic for CachePanic")
-		}()
-
-		// Test CachePanicWithHandle with nil
-		func() {
-			defer func() {
-				if r := recover(); r != nil {
-					t.Logf("Manual recovery caught panic with nil handle: %v", r)
-				}
-			}()
-			defer CachePanicWithHandle(nil)
-			panic("test panic with nil handle")
-		}()
-
-		// Test CachePanicWithHandle with custom handler
-		handleCalled := false
-		func() {
-			defer func() {
-				if r := recover(); r != nil {
-					t.Logf("Manual recovery in custom handler test: %v", r)
-				}
-			}()
-			defer CachePanicWithHandle(func(err interface{}) {
-				handleCalled = true
-				t.Logf("Custom handler called with: %v", err)
-			})
-			panic("test panic with handler")
-		}()
-		assert.True(t, handleCalled, "Custom panic handler should be called")
-
-		// Test with different panic types
-		panicTypes := []interface{}{
-			"string panic",
-			42,
-			[]string{"slice", "panic"},
-			struct{ msg string }{"struct panic"},
-			nil,
-		}
-
-		for i, panicValue := range panicTypes {
-			var capturedErr interface{}
-			func() {
-				defer func() {
-					if r := recover(); r != nil {
-						t.Logf("Manual recovery for panic type %d: %v", i, r)
-					}
-				}()
-				defer CachePanicWithHandle(func(err interface{}) {
-					capturedErr = err
-				})
-				panic(panicValue)
-			}()
-			assert.Equal(t, panicValue, capturedErr, "Should capture panic value correctly")
-		}
-	})
-
-	// Test concurrent access to all functions
-	t.Run("ConcurrentAccess", func(t *testing.T) {
-		const numGoroutines = 50
-		const numIterations = 10
-
-		var wg sync.WaitGroup
-		wg.Add(numGoroutines)
-
-		for i := 0; i < numGoroutines; i++ {
-			go func() {
-				defer wg.Done()
-				for j := 0; j < numIterations; j++ {
-					// Test all functions concurrently
-					ExecDir()
-					ExecFile()
-					Pwd()
-					UserHomeDir()
-					UserConfigDir()
-					UserCacheDir()
-					LazyConfigDir()
-					LazyCacheDir()
-					IsWindows()
-					IsDarwin()
-					IsLinux()
-					GetExitSign()
-					PrintStack()
-
-					// Test panic handling with manual recovery
-					func() {
-						defer func() {
-							if r := recover(); r != nil {
-								// Manual recovery for concurrent test
-							}
-						}()
-						defer CachePanic()
-						if j%5 == 0 {
-							panic("concurrent panic test")
-						}
-					}()
-				}
-			}()
-		}
-
-		wg.Wait()
-	})
-
-	// Test edge cases and error conditions
-	t.Run("EdgeCasesAndErrors", func(t *testing.T) {
-		// Test with modified working directory
-		originalDir, err := os.Getwd()
-		assert.NoError(t, err)
-		defer os.Chdir(originalDir)
-
-		// Test with root directory
-		if err := os.Chdir("/"); err == nil {
-			pwd := Pwd()
-			assert.NotEmpty(t, pwd)
-			execDir := ExecDir()
-			assert.NotEmpty(t, execDir)
-			execFile := ExecFile()
-			assert.NotEmpty(t, execFile)
-		}
-
-		// Test with temp directory
-		if err := os.Chdir(os.TempDir()); err == nil {
-			pwd := Pwd()
-			assert.NotEmpty(t, pwd)
-		}
-
-		// Test PrintStack with very small stack
-		debug.SetMaxStack(1)
-		PrintStack()
-		debug.SetMaxStack(8192) // Reset
-
-		// Test panic handling with empty stack scenario
-		func() {
-			defer func() {
-				if r := recover(); r != nil {
-					debug.SetMaxStack(1)
-					defer debug.SetMaxStack(8192)
-					PrintStack()
-					// Recovery handled manually
-				}
-			}()
-			defer CachePanic()
-			panic("empty stack test")
-		}()
-	})
-
-	// Test function return values validation
-	t.Run("ReturnValueValidation", func(t *testing.T) {
-		// All directory functions should return consistent values
-		execDir1 := ExecDir()
-		execDir2 := ExecDir()
-		assert.Equal(t, execDir1, execDir2, "ExecDir should return consistent values")
-
-		execFile1 := ExecFile()
-		execFile2 := ExecFile()
-		assert.Equal(t, execFile1, execFile2, "ExecFile should return consistent values")
-
-		pwd1 := Pwd()
-		pwd2 := Pwd()
-		assert.Equal(t, pwd1, pwd2, "Pwd should return consistent values")
-
-		// User directory functions should return consistent values
-		home1 := UserHomeDir()
-		home2 := UserHomeDir()
-		assert.Equal(t, home1, home2, "UserHomeDir should return consistent values")
-
-		config1 := UserConfigDir()
-		config2 := UserConfigDir()
-		assert.Equal(t, config1, config2, "UserConfigDir should return consistent values")
-
-		cache1 := UserCacheDir()
-		cache2 := UserCacheDir()
-		assert.Equal(t, cache1, cache2, "UserCacheDir should return consistent values")
-
-		lazyConfig1 := LazyConfigDir()
-		lazyConfig2 := LazyConfigDir()
-		assert.Equal(t, lazyConfig1, lazyConfig2, "LazyConfigDir should return consistent values")
-
-		lazyCache1 := LazyCacheDir()
-		lazyCache2 := LazyCacheDir()
-		assert.Equal(t, lazyCache1, lazyCache2, "LazyCacheDir should return consistent values")
 	})
 }
 
-// TestExitFunctionSafely tests Exit function in a safe way
-func TestExitFunctionSafely(t *testing.T) {
+// TestExitFunctionComponents tests only Exit function components safely
+func TestExitFunctionComponents(t *testing.T) {
 	t.Run("ExitFunctionComponents", func(t *testing.T) {
 		// Test FindProcess functionality (part of Exit function)
 		pid := os.Getpid()
@@ -365,47 +80,13 @@ func TestExitFunctionSafely(t *testing.T) {
 	})
 }
 
-// BenchmarkAllFunctions benchmarks all runtime functions
-func BenchmarkAllFunctions(b *testing.B) {
-	b.Run("ExecDir", func(b *testing.B) {
-		for i := 0; i < b.N; i++ {
-			ExecDir()
-		}
-	})
-
-	b.Run("ExecFile", func(b *testing.B) {
-		for i := 0; i < b.N; i++ {
-			ExecFile()
-		}
-	})
-
-	b.Run("Pwd", func(b *testing.B) {
-		for i := 0; i < b.N; i++ {
-			Pwd()
-		}
-	})
-
+// BenchmarkSystemFunctions benchmarks only system detection functions
+func BenchmarkSystemFunctions(b *testing.B) {
 	b.Run("SystemDetection", func(b *testing.B) {
 		for i := 0; i < b.N; i++ {
 			IsWindows()
 			IsDarwin()
 			IsLinux()
-		}
-	})
-
-	b.Run("UserDirectories", func(b *testing.B) {
-		for i := 0; i < b.N; i++ {
-			UserHomeDir()
-			UserConfigDir()
-			UserCacheDir()
-			LazyConfigDir()
-			LazyCacheDir()
-		}
-	})
-
-	b.Run("PrintStack", func(b *testing.B) {
-		for i := 0; i < b.N; i++ {
-			PrintStack()
 		}
 	})
