--- runtime/enhanced_coverage_test.go
@@ -1,288 +1,0 @@
-package runtime
-
-import (
-	"os"
-	"runtime/debug"
-	"strings"
-	"testing"
-)
-
-// TestEnhancedCoverage targets specific coverage gaps
-func TestEnhancedCoverage(t *testing.T) {
-	t.Run("cache_panic_with_handle_edge_cases", func(t *testing.T) {
-		// Test CachePanicWithHandle with empty stack scenario
-		defer func() {
-			if r := recover(); r != nil {
-				t.Logf("Recovered from test panic: %v", r)
-			}
-		}()
-
-		// Test with panic having empty stack case
-		func() {
-			defer CachePanicWithHandle(func(err interface{}) {
-				t.Logf("Handle called with: %v", err)
-			})
-			panic("test panic for coverage")
-		}()
-	})
-
-	t.Run("print_stack_edge_cases", func(t *testing.T) {
-		// Call PrintStack multiple times to ensure full coverage
-		PrintStack()
-
-		// Modify stack state and call again
-		func() {
-			PrintStack()
-		}()
-
-		// Test in different contexts
-		go func() {
-			PrintStack()
-		}()
-	})
-
-	t.Run("exec_dir_error_scenarios", func(t *testing.T) {
-		// Try to trigger the error path in ExecDir
-		// This is challenging because os.Executable() rarely fails in normal conditions
-
-		// Multiple calls to ensure all paths are covered
-		for i := 0; i < 5; i++ {
-			dir := ExecDir()
-			if dir == "" {
-				t.Logf("ExecDir returned empty on call %d (error path)", i)
-			} else {
-				t.Logf("ExecDir call %d: %s", i, dir)
-			}
-		}
-	})
-
-	t.Run("exec_file_error_scenarios", func(t *testing.T) {
-		// Try to trigger the error path in ExecFile
-		// Multiple calls to ensure all paths are covered
-		for i := 0; i < 5; i++ {
-			file := ExecFile()
-			if file == "" {
-				t.Logf("ExecFile returned empty on call %d (error path)", i)
-			} else {
-				t.Logf("ExecFile call %d: %s", i, file)
-			}
-		}
-	})
-
-	t.Run("pwd_error_scenarios", func(t *testing.T) {
-		// Try to trigger the error path in Pwd
-		// Multiple calls to ensure all paths are covered
-		for i := 0; i < 5; i++ {
-			pwd := Pwd()
-			if pwd == "" {
-				t.Logf("Pwd returned empty on call %d (error path)", i)
-			} else {
-				t.Logf("Pwd call %d: %s", i, pwd)
-			}
-		}
-	})
-
-	t.Run("stack_manipulation_coverage", func(t *testing.T) {
-		// Try to manipulate debug.Stack() behavior for coverage
-
-		// Call PrintStack in various contexts
-		defer func() {
-			if r := recover(); r != nil {
-				PrintStack() // Call during panic recovery
-			}
-		}()
-
-		// Normal call
-		PrintStack()
-
-		// Call in nested function
-		func() {
-			PrintStack()
-		}()
-
-		// Call with stack modification (safe values)
-		oldMaxStack := debug.SetMaxStack(2048)
-		PrintStack()
-		debug.SetMaxStack(oldMaxStack) // Restore original
-	})
-
-	t.Run("comprehensive_panic_scenarios", func(t *testing.T) {
-		// Test various panic scenarios to improve CachePanicWithHandle coverage
-
-		scenarios := []interface{}{
-			"string panic",
-			42,
-			struct{ msg string }{"structured panic"},
-			nil,
-		}
-
-		for i, scenario := range scenarios {
-			func() {
-				defer func() {
-					if r := recover(); r != nil {
-						t.Logf("Scenario %d recovered: %v", i, r)
-					}
-				}()
-
-				defer CachePanicWithHandle(func(err interface{}) {
-					t.Logf("Handle for scenario %d: %v", i, err)
-				})
-
-				panic(scenario)
-			}()
-		}
-	})
-
-	t.Run("stack_edge_cases", func(t *testing.T) {
-		// Test scenarios where stack might be empty or different
-
-		// Call in main goroutine
-		PrintStack()
-
-		// Call in new goroutine
-		done := make(chan bool, 1)
-		go func() {
-			defer func() { done <- true }()
-			PrintStack()
-		}()
-		<-done
-
-		// Call with modified stack size (safe value)
-		oldMaxStack := debug.SetMaxStack(4096)
-		PrintStack()
-		debug.SetMaxStack(oldMaxStack)
-	})
-}
-
-// TestCachePanicHandleVariations tests different handle function scenarios
-func TestCachePanicHandleVariations(t *testing.T) {
-	t.Run("handle_with_various_error_types", func(t *testing.T) {
-		errorTypes := []interface{}{
-			"simple string error",
-			42,
-			[]string{"slice", "error"},
-			map[string]interface{}{"complex": "error"},
-		}
-
-		for i, errType := range errorTypes {
-			func() {
-				defer func() {
-					if r := recover(); r != nil {
-						t.Logf("Test %d recovered: %v", i, r)
-					}
-				}()
-
-				defer CachePanicWithHandle(func(err interface{}) {
-					switch e := err.(type) {
-					case string:
-						t.Logf("Handle string error: %s", e)
-					case int:
-						t.Logf("Handle int error: %d", e)
-					case []string:
-						t.Logf("Handle slice error: %v", e)
-					case map[string]interface{}:
-						t.Logf("Handle map error: %v", e)
-					default:
-						t.Logf("Handle unknown error type: %v", e)
-					}
-				})
-
-				panic(errType)
-			}()
-		}
-	})
-
-	t.Run("handle_with_stack_logging", func(t *testing.T) {
-		// Test with handle that logs the stack
-		func() {
-			defer func() {
-				if r := recover(); r != nil {
-					t.Logf("Recovered after stack logging: %v", r)
-				}
-			}()
-
-			defer CachePanicWithHandle(func(err interface{}) {
-				t.Logf("Handle called with: %v", err)
-				// Also check the stack in the handle
-				stack := debug.Stack()
-				if len(stack) > 0 {
-					lines := strings.Split(string(stack), "\n")
-					t.Logf("Stack has %d lines", len(lines))
-				}
-			})
-
-			panic("stack logging test")
-		}()
-	})
-}
-
-// TestFilesystemFunctionStress performs stress testing to hit edge cases
-func TestFilesystemFunctionStress(t *testing.T) {
-	t.Run("concurrent_filesystem_calls", func(t *testing.T) {
-		const numGoroutines = 10
-		const callsPerGoroutine = 20
-
-		done := make(chan bool, numGoroutines)
-
-		for g := 0; g < numGoroutines; g++ {
-			go func(gid int) {
-				defer func() { done <- true }()
-
-				for i := 0; i < callsPerGoroutine; i++ {
-					// Call all filesystem functions
-					dir := ExecDir()
-					file := ExecFile()
-					pwd := Pwd()
-
-					// Log empty results (error paths)
-					if dir == "" {
-						t.Logf("G%d-I%d: ExecDir empty", gid, i)
-					}
-					if file == "" {
-						t.Logf("G%d-I%d: ExecFile empty", gid, i)
-					}
-					if pwd == "" {
-						t.Logf("G%d-I%d: Pwd empty", gid, i)
-					}
-				}
-			}(g)
-		}
-
-		for i := 0; i < numGoroutines; i++ {
-			<-done
-		}
-	})
-}
-
-// TestExitFunctionPrerequisites tests the components used by Exit()
-func TestExitFunctionPrerequisites(t *testing.T) {
-	t.Run("exit_function_components", func(t *testing.T) {
-		// Test os.FindProcess for current process
-		pid := os.Getpid()
-		process, err := os.FindProcess(pid)
-		if err != nil {
-			t.Logf("FindProcess error (testing error path): %v", err)
-		} else {
-			t.Logf("Found process: PID=%d", process.Pid)
-
-			// Test signal capability (without actually exiting)
-			// We can't call Exit() directly, but we can test its components
-			if process.Pid == pid {
-				t.Logf("Process PID matches current PID")
-			}
-		}
-	})
-
-	t.Run("signal_capability_test", func(t *testing.T) {
-		// Test the signal sending capability used by Exit()
-		process, err := os.FindProcess(os.Getpid())
-		if err != nil {
-			t.Logf("Signal test: FindProcess failed: %v", err)
-			return
-		}
-
-		// We test signal capability but don't actually send termination signals
-		// This validates the process.Signal path that Exit() would use
-		t.Logf("Signal capability available for PID: %d", process.Pid)
-	})
-}
