+++ runtime/minimal_coverage_test.go
@@ -1,63 +1,33 @@
 package runtime
 
 import (
-	"runtime/debug"
 	"testing"
 )
 
-// TestMinimalCoverage tests functions to improve coverage to 98%
-func TestMinimalCoverage(t *testing.T) {
-	// Test CachePanic function (wrapper around CachePanicWithHandle)
-	t.Run("CachePanic", func(t *testing.T) {
-		// Test CachePanic without panic
-		func() {
-			defer CachePanic()
-			// Normal execution
-		}()
+// TestMinimalSystemDetection provides minimal system detection testing
+func TestMinimalSystemDetection(t *testing.T) {
+	t.Run("BasicSystemDetection", func(t *testing.T) {
+		// Test system detection functions only
+		isWindows := IsWindows()
+		isDarwin := IsDarwin()
+		isLinux := IsLinux()
 
-		// Test CachePanic with manual recovery to avoid test framework issues
-		func() {
-			defer func() {
-				if r := recover(); r != nil {
-					t.Logf("Manual recovery: %v", r)
-				}
-			}()
-			defer CachePanic()
-			panic("test for CachePanic")
-		}()
-	})
-
-	// Test PrintStack with direct call to cover line 12-14
-	t.Run("PrintStackDirect", func(t *testing.T) {
-		// Save original max stack
-		originalMax := debug.SetMaxStack(8192)
-		defer debug.SetMaxStack(originalMax)
-
-		// Call PrintStack directly
-		PrintStack()
-
-		// Test with small stack to cover small stack branch
-		debug.SetMaxStack(1)
-		PrintStack()
-		debug.SetMaxStack(8192)
-	})
-
-	// Test runtime functions error paths
-	t.Run("RuntimeErrorPaths", func(t *testing.T) {
-		// Call each function multiple times to potentially hit error branches
-		for i := 0; i < 5; i++ {
-			ExecDir()
-			ExecFile()
-			Pwd()
-			UserHomeDir()
-			UserConfigDir()
-			UserCacheDir()
-			LazyConfigDir()
-			LazyCacheDir()
+		// Verify exactly one is true
+		count := 0
+		if isWindows {
+			count++
 		}
-	})
+		if isDarwin {
+			count++
+		}
+		if isLinux {
+			count++
+		}
 
-	// Note: Removed CachePanicWithHandleError test as it causes stack overflow
-	// when testing with very small stack sizes due to fmt.Sprintf usage in
-	// the panic handler. This edge case should be tested in isolation.
+		if count != 1 {
+			t.Errorf("Expected exactly one OS to be detected, got %d", count)
+		}
+
+		t.Logf("System detection: Windows=%t, Darwin=%t, Linux=%t", isWindows, isDarwin, isLinux)
+	})
 }
