--- runtime/missing_coverage_test.go
@@ -1,245 +1,0 @@
-package runtime
-
-import (
-	"os"
-	"runtime/debug"
-	"strings"
-	"testing"
-)
-
-// TestMissingCoverage tests uncovered code paths to improve coverage
-func TestMissingCoverage(t *testing.T) {
-	t.Run("PrintStack_empty_stack_case", func(t *testing.T) {
-		// Test PrintStack multiple times to ensure all branches are covered
-		PrintStack()
-		PrintStack()
-		t.Log("PrintStack called successfully")
-	})
-
-	t.Run("CachePanicWithHandle_empty_stack_branch", func(t *testing.T) {
-		// Test the branch where stack might be empty
-		defer func() {
-			if r := recover(); r != nil {
-				t.Logf("Recovered in test: %v", r)
-			}
-		}()
-
-		func() {
-			defer CachePanicWithHandle(func(err interface{}) {
-				t.Logf("Handle called with: %v", err)
-			})
-			panic("test panic for stack coverage")
-		}()
-	})
-
-	t.Run("ExecDir_error_path", func(t *testing.T) {
-		// We can't easily trigger os.Executable() error,
-		// but we can test the function multiple times
-		dir1 := ExecDir()
-		dir2 := ExecDir()
-
-		// Both calls should return the same directory
-		if dir1 != dir2 {
-			t.Errorf("ExecDir should return consistent results: %s vs %s", dir1, dir2)
-		}
-
-		if dir1 == "" {
-			t.Log("ExecDir returned empty (error path covered)")
-		} else {
-			t.Logf("ExecDir returned: %s", dir1)
-		}
-	})
-
-	t.Run("ExecFile_error_path", func(t *testing.T) {
-		// Similar to ExecDir
-		file1 := ExecFile()
-		file2 := ExecFile()
-
-		if file1 != file2 {
-			t.Errorf("ExecFile should return consistent results: %s vs %s", file1, file2)
-		}
-
-		if file1 == "" {
-			t.Log("ExecFile returned empty (error path covered)")
-		} else {
-			t.Logf("ExecFile returned: %s", file1)
-		}
-	})
-
-	t.Run("Pwd_error_path", func(t *testing.T) {
-		// Similar to other path functions
-		pwd1 := Pwd()
-		pwd2 := Pwd()
-
-		if pwd1 != pwd2 {
-			t.Errorf("Pwd should return consistent results: %s vs %s", pwd1, pwd2)
-		}
-
-		if pwd1 == "" {
-			t.Log("Pwd returned empty (error path covered)")
-		} else {
-			t.Logf("Pwd returned: %s", pwd1)
-		}
-	})
-
-	t.Run("comprehensive_panic_testing", func(t *testing.T) {
-		// Test various panic scenarios to improve coverage
-		testCases := []string{
-			"string panic",
-			"",
-		}
-
-		for i, testCase := range testCases {
-			func() {
-				defer func() {
-					if r := recover(); r != nil {
-						t.Logf("Test case %d recovered: %v", i, r)
-					}
-				}()
-
-				func() {
-					defer CachePanicWithHandle(func(err interface{}) {
-						t.Logf("Handle for test case %d: %v", i, err)
-					})
-					if testCase != "" {
-						panic(testCase)
-					}
-				}()
-			}()
-		}
-	})
-
-	t.Run("stack_analysis", func(t *testing.T) {
-		// Test stack analysis to improve coverage of stack-related code
-		st := debug.Stack()
-		if len(st) > 0 {
-			lines := strings.Split(string(st), "\n")
-			t.Logf("Stack has %d lines", len(lines))
-
-			// This exercises the loop in CachePanicWithHandle
-			for i, line := range lines {
-				if i < 3 { // Only log first few lines
-					t.Logf("Stack line %d: %s", i, line)
-				}
-			}
-		}
-	})
-}
-
-// TestExitFunctionComponents tests components used by Exit function
-func TestExitFunctionComponents(t *testing.T) {
-	t.Run("find_process_current", func(t *testing.T) {
-		// Test finding current process (used by Exit)
-		process, err := os.FindProcess(os.Getpid())
-		if err != nil {
-			t.Errorf("Failed to find current process: %v", err)
-		}
-
-		if process.Pid != os.Getpid() {
-			t.Errorf("Process PID mismatch: expected %d, got %d", os.Getpid(), process.Pid)
-		}
-
-		t.Logf("Current process found: PID %d", process.Pid)
-	})
-
-	t.Run("process_signal_capability", func(t *testing.T) {
-		// Test the capability to send signals (used by Exit)
-		process, err := os.FindProcess(os.Getpid())
-		if err != nil {
-			t.Fatalf("Failed to find current process: %v", err)
-		}
-
-		// Test sending a signal (this won't terminate the process)
-		err = process.Signal(os.Interrupt)
-		if err != nil {
-			t.Logf("Signal sending failed: %v (may be expected on some systems)", err)
-		} else {
-			t.Log("Signal capability confirmed")
-		}
-	})
-}
-
-// TestErrorSimulation tries to simulate error conditions
-func TestErrorSimulation(t *testing.T) {
-	t.Run("invalid_process_handling", func(t *testing.T) {
-		// Test handling of invalid process (error path in Exit)
-		invalidPid := 999999
-		process, err := os.FindProcess(invalidPid)
-
-		if err != nil {
-			t.Logf("FindProcess failed for invalid PID (expected): %v", err)
-		} else if process != nil {
-			// On some systems, FindProcess doesn't fail immediately
-			err = process.Signal(os.Interrupt)
-			if err != nil {
-				t.Logf("Signal to invalid process failed (expected): %v", err)
-			}
-		}
-	})
-}
-
-// TestAllFunctionEdgeCases tests edge cases for all functions
-func TestAllFunctionEdgeCases(t *testing.T) {
-	t.Run("all_functions_multiple_calls", func(t *testing.T) {
-		// Call all functions multiple times to improve coverage
-		const iterations = 5
-
-		for i := 0; i < iterations; i++ {
-			// System detection
-			_ = IsWindows()
-			_ = IsDarwin()
-			_ = IsLinux()
-
-			// Path functions
-			_ = ExecDir()
-			_ = ExecFile()
-			_ = Pwd()
-
-			// User directories
-			_ = UserHomeDir()
-			_ = UserConfigDir()
-			_ = UserCacheDir()
-			_ = LazyConfigDir()
-			_ = LazyCacheDir()
-
-			// Stack and panic functions
-			PrintStack()
-
-			func() {
-				defer CachePanic()
-			}()
-
-			func() {
-				defer CachePanicWithHandle(nil)
-			}()
-		}
-
-		t.Logf("Completed %d iterations of all function calls", iterations)
-	})
-}
-
-// TestBranchCoverage specifically targets missed branches
-func TestBranchCoverage(t *testing.T) {
-	t.Run("empty_string_paths", func(t *testing.T) {
-		// Test when path functions might return empty strings
-		funcs := map[string]func() string{
-			"ExecDir":       ExecDir,
-			"ExecFile":      ExecFile,
-			"Pwd":           Pwd,
-			"UserHomeDir":   UserHomeDir,
-			"UserConfigDir": UserConfigDir,
-			"UserCacheDir":  UserCacheDir,
-			"LazyConfigDir": LazyConfigDir,
-			"LazyCacheDir":  LazyCacheDir,
-		}
-
-		for name, fn := range funcs {
-			result := fn()
-			if result == "" {
-				t.Logf("%s returned empty string (error path)", name)
-			} else {
-				t.Logf("%s returned: %s", name, result)
-			}
-		}
-	})
-}
